+++ util/options_helper.cc
@@ -391,6 +391,8 @@ bool ParseColumnFamilyOption(const std::string& name, const std::string& value,
       new_options->table_factory.reset(NewBlockBasedTableFactory(table_opt));
     } else if (name == "min_write_buffer_number_to_merge") {
       new_options->min_write_buffer_number_to_merge = ParseInt(value);
+    } else if (name == "max_write_buffer_number_to_maintain") {
+      new_options->max_write_buffer_number_to_maintain = ParseInt(value);
     } else if (name == "compression") {
       new_options->compression = ParseCompressionType(value);
     } else if (name == "compression_per_level") {